fix(mongo): Delete claimed message fails
This patch adds an additional check on the primary node for the correct
claim ID if it fails to match on the secondary. This takes care of the
case when the claimed version of the message document has not yet been
replicated to the secondary before the client attempts to delete it.

Change-Id: Ia22562532351a36bedc980c7c6041517207b2723
Closes-Bug: #1247297